+++ Portierniaktosiedzi/Models/Employee.cs
@@ -7,7 +7,7 @@ namespace Portierniaktosiedzi.Models
 {
     public class Employee
     {
-        public Employee(float posts, Gender gender, string name)
+        public Employee(decimal posts, Gender gender, string name)
         {
             Posts = posts;
             Gender = gender;
@@ -18,7 +18,7 @@ namespace Portierniaktosiedzi.Models
 
         public Gender Gender { get; }
 
-        public float Posts { get; }
+        public decimal Posts { get; }
 
         public decimal GetWorkingHours(int month, int year, IHolidays holidays)
         {
@@ -34,7 +34,7 @@ namespace Portierniaktosiedzi.Models
                     }
                 }
 
-            return (decimal)Posts * workingdays * 8;
+            return Posts * workingdays * 8;
         }
     }
 }
